requirements.txt
# Core ML/NLP
datasets>=2.14.0
pandas>=1.5.0
numpy>=1.24.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Qdrant
qdrant-client>=1.7.0
python-dotenv>=1.0.1
# Utils
tqdm>=4.65.0

cloudinary>=1.44.1